back/requirements.txt
pymongo==4.1.1
redis==4.3.4
Flask-RESTful==0.3.9
Flask-Cors==3.0.10
python-dotenv==0.20.0